+++ doc/mf6io/mf6ivar/dfn/sim-nam.dfn
@@ -51,12 +51,12 @@ longname print input to listing file
 description keyword to activate printing of simulation input summaries to the simulation list file (mfsim.lst). With this keyword, input summaries will be written for those packages that support newer input data model routines.  Not all packages are supported yet by the newer input data model routines.
 
 block options
-name tolerate_unknown
+name strict
 type keyword
 reader urword
 optional true
-longname tolerate unknown options
-description keyword to indicate that unrecognized top-level input option tags should be treated as warnings rather than errors. When this keyword is specified, any top-level tag (not a subfield of a record) that is not recognized for a package will generate a warning and be skipped. This allows input files written for a newer version of MODFLOW 6 to be used with an older version, provided the unrecognized options are not required.
+longname strict input checking
+description keyword to indicate that unrecognized top-level input option tags should be treated as errors rather than warnings. By default, any top-level tag (not a subfield of a record) that is not recognized for a package will generate a warning and be skipped, allowing input files written for a newer version of MODFLOW 6 to be used with an older version. When STRICT is specified, unrecognized tags will instead cause the simulation to terminate with an error.
 
 block options
 name hpc_filerecord
