+++ src/Utilities/SimVariables.f90
@@ -33,7 +33,7 @@ module SimVariablesModule
   integer(I4B) :: iout !< file unit number for simulation output
   integer(I4B) :: isimcnvg !< simulation convergence flag (1) if all objects have converged, (0) otherwise
   integer(I4B) :: isimcontinue = 0 !< simulation continue flag (1) to continue if isimcnvg = 0, (0) to terminate
-  integer(I4B) :: itolerate_unknown = 0 !< tolerate unknown input options (1) warn and skip, (0) error
+  integer(I4B) :: isimstrict = 0 !< strict input checking (1) error on unknown options, (0) warn and skip
   integer(I4B) :: nocheck = 0 !< nocheck option (0) to check input, (1) to ignore checks
   integer(I4B) :: isimcheck = 1 !< simulation input check flag (1) to check input, (0) to ignore checks
   integer(I4B) :: numnoconverge = 0 !< number of times the simulation did not converge
